code split TaskAssignments
