+++ project/assets/mgmt/TaskAssignments/index.js
@@ -1,13 +1,11 @@
-import React from 'react';
-import { render } from 'react-dom';
-import { Provider } from 'react-redux';
-
-import Root from './containers/Root';
-import configureStore from './store/configureStore';
+import { splitStartupRedux } from 'utils/WebpackSplit';
 
 const startup = function(element){
-    const store = configureStore();
-    render(<Provider store={store}><Root /></Provider>, element);
+    import('mgmt/TaskAssignments/containers/Root').then((Component) => {
+        import('mgmt/TaskAssignments/store/configureStore').then((store) => {
+            splitStartupRedux(element, Component.default, store.default);
+        });
+    });
 };
 
 export default startup;
